mcp-server/pyproject.toml
[project]
name = "mcp-server"
version = "0.1.0"
description = "Semantic skill discovery backend using ChromaDB and MCP protocol"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "anyio>=4.11.0",
    "chromadb>=1.3.5",
    "mcp>=0.1.0",
    "pyyaml>=6.0.3",
]

[project.scripts]
mcp-serve = "mcp_server:main"
mcp-index = "index_skills:main"
